implementado validação de conexao com internet em todo o app

+++ app/src/main/java/com/ahpp/notshoes/util/cards/CardListaDesejos.kt
@@ -1,5 +1,6 @@
 package com.ahpp.notshoes.util.cards
 
+import android.widget.Toast
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -43,6 +44,7 @@ import com.ahpp.notshoes.R
 import com.ahpp.notshoes.bd.produto.ProdutoRepository
 import com.ahpp.notshoes.model.Produto
 import com.ahpp.notshoes.util.clienteLogado
+import com.ahpp.notshoes.util.funcoes.possuiConexao
 import com.ahpp.notshoes.util.produtoSelecionado
 import java.text.NumberFormat
 
@@ -52,13 +54,14 @@ fun CardListaDesejos(
     produto: Produto,
     onRemoveProduct: (Produto) -> Unit
 ) {
+    val ctx = LocalContext.current
 
     val localeBR = java.util.Locale("pt", "BR")
     val numberFormat = NumberFormat.getCurrencyInstance(localeBR)
 
     //imagem do produto
     val painter = rememberAsyncImagePainter(
-        model = ImageRequest.Builder(LocalContext.current)
+        model = ImageRequest.Builder(ctx)
             .data(produto.imagemProduto)
             .crossfade(true)
             .size(Size.ORIGINAL)
@@ -168,12 +171,16 @@ fun CardListaDesejos(
                 Button(
                     modifier = Modifier.size(30.dp), contentPadding = PaddingValues(0.dp),
                     onClick = {
-                        val repository = ProdutoRepository()
-                        repository.removerProdutoListaDesejos(
-                            produto.idProduto,
-                            clienteLogado.idCliente
-                        )
-                        onRemoveProduct(produto)
+                        if (possuiConexao(ctx)) {
+                            val repository = ProdutoRepository()
+                            repository.removerProdutoListaDesejos(
+                                produto.idProduto,
+                                clienteLogado.idCliente
+                            )
+                            onRemoveProduct(produto)
+                        } else {
+                            Toast.makeText(ctx, "Erro de rede.", Toast.LENGTH_SHORT).show()
+                        }
                     },
                     colors = ButtonDefaults.buttonColors(Color.White),
                     elevation = ButtonDefaults.buttonElevation(10.dp)
